+++ index.js
@@ -20,6 +20,11 @@ async function run() {
         let rulesFileLocation = core.getInput('rules_file_name');
         let cmdOptions = core.getInput('cmd_options');
         let issueTitle = core.getInput('issue_title');
+        let failAction = core.getInput('fail_action');
+
+        if (!(String(failAction).toLowerCase() === 'true' || String(failAction).toLowerCase() === 'false')) {
+            console.log('[WARNING]: \'fail_action\' action input should be either \'true\' or \'false\'');
+        }
 
         console.log('starting the program');
         console.log('github run id :' + currentRunnerID);
@@ -29,6 +34,7 @@ async function run() {
             plugins = await common.helper.processLineByLine(`${workspace}/${rulesFileLocation}`);
         }
 
+        await exec.exec(`docker pull ${docker_name} -q`);
         let command = (`docker run --user root -v ${workspace}:/zap/wrk/:rw --network="host" ` +
             `-t ${docker_name} zap-baseline.py -t ${target} -J ${jsonReportName} -w ${mdReportName}  -r ${htmlReportName} ${cmdOptions}`);
 
@@ -39,7 +45,18 @@ async function run() {
         try {
             await exec.exec(command);
         } catch (err) {
-            core.setFailed('The ZAP Baseline scan has failed, starting to analyze the alerts. err: ' + err.toString());
+            if (err.toString().includes('exit code 3')) {
+                core.setFailed('failed to scan the target: ' + err.toString());
+                return
+            }
+
+            if ((err.toString().includes('exit code 2') || err.toString().includes('exit code 1'))
+                    && String(failAction).toLowerCase() === 'true') {
+                console.log(`[info] By default ZAP Docker container will fail if it identifies any alerts during the scan!`);
+                core.setFailed('Scan action failed as ZAP has identified alerts, starting to analyze the results. ' + err.toString());
+            }else {
+                console.log('Scanning process completed, starting to analyze the results!')
+            }
         }
         await common.main.processReport(token, workspace, plugins, currentRunnerID, issueTitle, repoName);
     } catch (error) {